+++ test/controllers/configuration/exports_controller_test.rb
@@ -118,6 +118,28 @@ class ExportsControllerTest < ActionDispatch::IntegrationTest
     assert_match dossier.eleve.identifiant, xml.css("ID_NATIONAL").text
   end
 
+  test "Pour une naissance à l'étranger, la commune est renseignée sous forme de texte" do
+    admin = Fabricate(:admin)
+    identification_agent(admin)
+
+    eleve = Fabricate(:eleve, ville_naiss: "KINSHASA", pays_naiss: "324")
+    dossier = Fabricate(:dossier_eleve,
+                        eleve: eleve,
+                        etablissement: admin.etablissement,
+                        mef_destination: Fabricate(:mef, etablissement: admin.etablissement))
+
+    get export_siecle_configuration_exports_path(xml_only: true)
+
+    assert_response :success
+
+    schema = Rails.root.join("doc/import_prive/schema_Import_3.1.xsd")
+    Nokogiri::XML::Schema(File.read(schema))
+    xml = Nokogiri::XML(response.body)
+
+    assert_match dossier.eleve.pays_naiss,  xml.css("CODE_PAYS").text
+    assert_match dossier.eleve.ville_naiss, xml.css("VILLE_NAISS").text
+  end
+
   test "export uniquement pour l'INE saisi" do
     admin = Fabricate(:admin)
     identification_agent(admin)