+++ test/fabricators/eleve_fabricator.rb
@@ -6,4 +6,5 @@ Fabricator(:eleve) do
   identifiant { sequence(:identifiant) { |i| "IDENTIF#{'I' * (4 - i.to_s.size)}#{i}" } }
   date_naiss "2004-04-27"
   commune_insee_naissance "75112"
+  pays_naiss "100"
 end